+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.0.1
+## 1.0.0
 
-* TODO: Describe initial release.
+- Initial release.
+- Added DOT Code parsing and validation.
+- Added support for Load Index parsing, validation, and conversion between lbs and kg.
+- Added support for Speed Index parsing, validation, and conversion between km/h and mph.